refactor: arrange folder structure for addition of sub hospitals feature
