--- src/routes/dashboard/landing/hospitalCapacity/index.ts
@@ -1,3 +1,0 @@
-import HospitalCapacity from './HospitalCapacity';
-
-export default HospitalCapacity;
