+++ src/routes/dashboard/landing/hospitals/hospitalCapacity/index.ts
@@ -1,0 +1,3 @@
+import HospitalCapacity from "./HospitalCapacity";
+
+export default HospitalCapacity;
