+++ src/routes/dashboard/landing/hospitals/index.ts
@@ -1,0 +1,3 @@
+import Hospitals from "./Hospitals";
+
+export default Hospitals;
